+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## WIP
+## 0.4.5 - 2026-03-24
 
 ### Features